[fix] ver1.0 build:2 リジェクト原因のタイトル名を修正　→　再提出
Design - 4.1

Your app or its metadata contains misleading content. Specifically, your app creates a misleading association with Pokemon￼.
Next Steps
Please remove or revise any misleading content in your app and its metadata.
In addition, we recommend adjusting your app name so that the misleading element is used as a descriptor, not as part of the product name. For example, the following style is an acceptable format for names:
PutYourUniqueAppNameHere - with tagline "for Pokemon￼”